Remove sourcing of virtual environment scripts
Systemd expects a path to a executable as the argument to ExecStartPre. Sourcing the activate script assumes that the commands are run within a shell, which they are not. The document as it is currently written is just plain wrong.

Fortunately there is no need to set up the virtual environment as python tools installed within one are aware of how they were installed, and have their sys.path set correctly automatically. Just executing `hass` directly is enough.